dashboardApp/requirements.txt
Flask==2.3.2
Flask-Cors==5.0.0
google-cloud-firestore==2.19.0
gunicorn==23.0.0
python-dotenv==1.0.1